Move LocationSearch
