+++ packages/client/src/views/SysAdmin/Performance/utils.tsx
@@ -22,7 +22,7 @@ import endOfMonth from 'date-fns/endOfMonth'
 import React from 'react'
 import { getPercentage } from '@client/utils/data-formatting'
 import { FormattedNumber, IntlShape } from 'react-intl'
-import { ISearchLocation } from '@opencrvs/components/lib/interface'
+import { ISearchLocation } from '@opencrvs/components/lib/LocationSearch'
 import { ListViewSimplified } from '@opencrvs/components/lib/ListViewSimplified'
 import { constantsMessages } from '@client/i18n/messages'
 import { messages as statusMessages } from '@client/i18n/messages/views/registrarHome'
